Fixed 64bit integers on the stack, again, attempting to fix issue with returning from Mernel now

+++ Nernel/Include/CLR/InternalCalls/System.DateTime.h
@@ -2,4 +2,4 @@
 
 #include <CLR/AppDomain.h>
 
-uint32_t System_DateTime_InternalUtcNow(AppDomain* pAppDomain);
+uint64_t System_DateTime_InternalUtcNow(AppDomain* pAppDomain);
